Total income of an owner includes defined benefits from a scale class.

+++ FishingBooker/Models/LoyaltyScaleViewModel.cs
@@ -20,17 +20,17 @@ namespace FishingBooker.Models
 
         [Display(Name = "Clients benefits for this scale(%)")]
         [Range(1, 100, ErrorMessage = "Percentage must be between 1 and 100!")]
-        public decimal ClientsBenefits { get; set; }
+        public float ClientsBenefits { get; set; }
 
 
         [Display(Name = "Owners benefits for this scale(%)")]
         [Range(1, 100, ErrorMessage = "Percentage must be between 1 and 100!")]
-        public decimal OwnerBenefits { get; set; }
+        public float OwnerBenefits { get; set; }
 
 
         [Display(Name = "Minumum earned points for a scale")]
         [Range(0, 999999, ErrorMessage = "Range is between 0 and 999999")]
         [Required(ErrorMessage = "You need to enter a number.")]
-        public decimal MinEarnedPoints { get; set; }
+        public float MinEarnedPoints { get; set; }
     }
 }
